move collection types internally and export only public ones
