+++ integration/test_collection.py
@@ -5,7 +5,7 @@ import pytest as pytest
 import weaviate
 from weaviate import Config
 from weaviate.collection.grpc import HybridFusion, LinkTo, MetadataQuery
-from weaviate.weaviate_classes import (
+from weaviate.collection.classes import (
     CollectionConfig,
     Property,
     DataType,
